# backend-api/api/service/openapi.yaml
openapi: 3.0.3
info:
  title: WAS API
  version: "1.0.0"
...
paths:
...
  /api/v1/security-profiles:
    get:
      operationId: ListSecurityProfiles
      tags:
        - security-profiles
      summary: List security profiles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityProfile"
    post:
      operationId: CreateSecurityProfile
      tags:
        - security-profiles
      summary: Create security profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityProfileCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityProfile"
  /api/v1/security-profiles/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetSecurityProfile
      tags:
        - security-profiles
      summary: Get security profile by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityProfile"
        "404":
          description: Not found
    put:
      operationId: UpdateSecurityProfile
      tags:
        - security-profiles
      summary: Update security profile by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityProfile"
        "404":
          description: Not found
    delete:
      operationId: DeleteSecurityProfile
      tags:
        - security-profiles
      summary: Delete security profile by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
  /api/v1/security-rules:
    get:
      operationId: ListSecurityRules
      tags:
        - security-rules
      summary: List security rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityRule"
    post:
      operationId: CreateSecurityRule
      tags:
        - security-rules
      summary: Create security rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityRuleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityRule"
  /api/v1/security-rules/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetSecurityRule
      tags:
        - security-rules
      summary: Get security rule by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityRule"
        "404":
          description: Not found
    put:
      operationId: UpdateSecurityRule
      tags:
        - security-rules
      summary: Update security rule by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityRuleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityRule"
        "404":
          description: Not found
    delete:
      operationId: DeleteSecurityRule
      tags:
        - security-rules
      summary: Delete security rule by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
...
components:
  schemas:
...
    SecurityProfile:
      type: object
      required:
        - id
        - name
        - base_action
        - log_enabled
        - is_enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        base_action:
          type: string
          enum: [allow, block]
        log_enabled:
          type: boolean
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SecurityProfileCreate:
      type: object
      required:
        - name
        - base_action
        - log_enabled
        - is_enabled
      properties:
        name:
          type: string
        description:
          type: string
        base_action:
          type: string
          enum: [allow, block]
        log_enabled:
          type: boolean
        is_enabled:
          type: boolean
    SecurityProfileUpdate:
      type: object
      required:
        - name
        - base_action
        - log_enabled
        - is_enabled
      properties:
        name:
          type: string
        description:
          type: string
        base_action:
          type: string
          enum: [allow, block]
        log_enabled:
          type: boolean
        is_enabled:
          type: boolean
    SecurityRule:
      type: object
      required:
        - id
        - profile_id
        - name
        - priority
        - rule_type
        - action
        - dry_run
        - is_enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        rule_type:
          type: string
          enum: [deterministic, ml]
        action:
          type: string
          enum: [allow, block]
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        ml_model_id:
          type: integer
          format: int64
          nullable: true
        ml_threshold:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          nullable: true
        dry_run:
          type: boolean
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SecurityRuleCreate:
      type: object
      required:
        - profile_id
        - name
        - priority
        - rule_type
        - action
        - dry_run
        - is_enabled
      properties:
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        rule_type:
          type: string
          enum: [deterministic, ml]
        action:
          type: string
          enum: [allow, block]
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        ml_model_id:
          type: integer
          format: int64
          nullable: true
        ml_threshold:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          nullable: true
        dry_run:
          type: boolean
        is_enabled:
          type: boolean
    SecurityRuleUpdate:
      type: object
      required:
        - profile_id
        - name
        - priority
        - rule_type
        - action
        - dry_run
        - is_enabled
      properties:
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        rule_type:
          type: string
          enum: [deterministic, ml]
        action:
          type: string
          enum: [allow, block]
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        ml_model_id:
          type: integer
          format: int64
          nullable: true
        ml_threshold:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          nullable: true
        dry_run:
          type: boolean
        is_enabled:
          type: boolean
...
    SecurityHeaderCondition:
      type: object
      required:
        - name
        - value_regex
      properties:
        name:
          type: string
        value_regex:
          type: array
          items:
            type: string
    SecurityRuleConditions:
      type: object
      description: |
        Условия для security rules. Блоки разных типов объединяются по AND.
        Внутри одного блока значения объединяются по OR.
      properties:
        source_ip_cidr:
          type: array
          items:
            type: string
          description: CIDR диапазоны клиентских IP
        uri_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для HTTP URI
        host_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для host
        method_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для HTTP метода
        headers:
          type: array
          items:
            $ref: "#/components/schemas/SecurityHeaderCondition"
          description: Условия по имени заголовка и regexp значения
...